# Performance improvements for dasherize.