@rafaelfranca 2017-10-23
> params[:foo].dasherize → user input leads to growing cache size until it exhausts all memory, causing denial of service?

Yes.